+++ client/package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "axios": "^1.1.3",
     "bootstrap": "^5.2.3",
+    "bootstrap-icons": "^1.10.3",
     "react": "^17.0.1",
     "react-bootstrap": "^2.7.0",
     "react-dom": "^17.0.2",
