+++ client/src/App.js
@@ -4,8 +4,8 @@ import Home from './components/Home/Home'
 import 'bootstrap/dist/css/bootstrap.min.css';
 import Landing from './components/Landing';
 import Join from './components/Join'
-import Shop from './components/Shop'
-import Detail from './components/Detail/Detail';
+/* import Shop from './components/Shop' */
+import Detail from './components/Detail';
 
 function App() {
   return (
@@ -14,7 +14,7 @@ function App() {
         <Route exact path={'/'} component={Detail} />
         <Route path={'/home'} component={Home} />
         <Route path={'/join'} component={Join} />
-        <Route path={'/shop'} component={Shop} />
+        {/* <Route path={'/shop'} component={Shop} /> */}
         <Route path={'/shop/:id'} component={Detail} />
       </Switch>
     </BrowserRouter>
