+++ client/src/index.js
@@ -3,6 +3,7 @@ import ReactDOM from "react-dom";
 import App from "./App";
 import { Provider } from "react-redux";
 import { store } from "./store";
+import 'bootstrap-icons/font/bootstrap-icons.css';
 
 ReactDOM.render(
   <Provider store={store}>
